+++ R/features.R
@@ -244,9 +244,9 @@ setMethod(
       x@n == y@n,
       all(method %in% c("chisquare", "ll"))
     )
-    tokenColnames <- sapply(1:x@n, function(i) paste(i, x@p_attribute, sep = "_"))
+    token_colnames <- sapply(1L:x@n, function(i) paste(x@p_attribute, i, sep = "_"))
     z <- callNextMethod(
-      x = x, y = y, by = tokenColnames,
+      x = x, y = y, by = token_colnames,
       included = included, method = method, verbose = verbose
     )
     z <- as(z, "features_ngrams")